Update core.py to TRNSYS 18 + custom file formats to use
This commit updates the library to use TRNSYS18 and it gives the chance to copy custom file formats (that the user can decide) over to the batch files to be used in the simulation. Like python and matlab files. In the case of matlab it won't work with parallelization but i'll see if i can automatize a 'False' to parallelization in the case matlab files are imported.